+++ src/index.ts
@@ -83,6 +83,9 @@ export type { DatePickerProps, DatePickerInvalidReason } from './components/Date
 export { Dialog } from './components/Dialog/index';
 export type { DialogProps, DialogSize } from './components/Dialog/index';
 
+export { Tag, tagSizes } from './components/Tag/index';
+export type { TagProps, TagSize } from './components/Tag/index';
+
 export { Tabs, tabsVariants } from './components/Tabs/index';
 export type { TabsProps, TabItem, TabsVariant } from './components/Tabs/index';